+++ src/ext_widgets/scroll_view/children_layouter_list_view.c
@@ -202,41 +202,186 @@ static ret_t children_layouter_list_view_get_param(children_layouter_t* layouter
   return RET_FAIL;
 }
 
-static ret_t children_layouter_for_list_view_layout(children_layouter_t* layouter,
-                                                    widget_t* widget) {
-  int32_t spacing = 0;
-  int32_t x_margin = 0;
-  int32_t y_margin = 0;
+static ret_t children_layouter_list_view_for_list_view_children_layout_h(darray_t* children_for_layout, int32_t item_height, int32_t default_item_height) {
+  int32_t i = 0;
+  int32_t h = 0;
+  int32_t n = 0;
+  widget_t** children = NULL;
+  return_value_if_fail(children_for_layout != NULL, RET_BAD_PARAMS);
+  n = children_for_layout->size;
+  children = (widget_t**)children_for_layout->elms;
+  for (i = 0; i < n; i++) {
+    widget_t* iter = children[i];
+    if (iter->self_layout != NULL) {
+        widget_layout_self(iter);
+    }
+    if (iter->auto_adjust_size) {
+      widget_auto_adjust_size(iter);
+    }
+
+    h = item_height;
+    if (h <= 0) {
+      h = iter->h;
+    }
+    if (h <= 0) {
+      h = default_item_height;
+    }
+    widget_resize(iter, iter->w, h);
+  }
+  return RET_OK;
+}
+
+static ret_t children_layouter_list_view_for_list_view_children_layout_w(darray_t* children_for_layout, uint32_t cols, int32_t x_margin, int32_t y_margin, int32_t spacing, int32_t scroll_view_w) {
+  int32_t i = 0;
+  int32_t w = 0;
+  int32_t x = x_margin;
+  int32_t y = y_margin;
+  widget_t** children = NULL;
+  return_value_if_fail(children_for_layout != NULL, RET_BAD_PARAMS);
+  w = scroll_view_w - 2 * x_margin;
+  children = (widget_t**)children_for_layout->elms;
+  if (cols <= 1) {
+    for (i = 0; i < children_for_layout->size; i++) {
+      widget_t* iter = children[i];
+      widget_move_resize(iter, x, y, w, iter->h);
+      widget_layout_children(iter);
+      y += (iter->h + spacing);
+    }
+  } else {
+    int32_t j = 0;
+    int32_t n = 0;
+    int32_t h = 0;
+    int32_t size = children_for_layout->size;
+    uint32_t rows = (size % cols) ? (size / cols) + 1 : (size / cols);
+
+    w = (w - (cols - 1) * spacing) / cols;
+    for (i = 0; i < rows && n < size; i++) {
+      h = 0;
+      for (j = 0; j < cols && n < size; j++, n++) {
+        widget_t* iter = children[n];
+        int32_t tmp_x = x + j * (w + spacing);
+        widget_move_resize(iter, tmp_x, y, w, iter->h);
+        widget_layout_children(iter);
+        h = tk_max(h, iter->h);
+      }
+      y += (h + spacing);
+    }
+  }
+  return RET_OK;
+}
+
+static int32_t children_layouter_list_view_for_list_view_get_virtual_h(darray_t* children_for_layout, uint32_t cols, int32_t y_margin, int32_t spacing) {
+  int32_t i = 0;
+  widget_t** children = NULL;
+  int32_t virtual_h = y_margin;
+  return_value_if_fail(children_for_layout != NULL, 0);
+  children = (widget_t**)children_for_layout->elms;
+  if (cols <= 1) {
+    for (i = 0; i < children_for_layout->size; i++) {
+      virtual_h += (children[i]->h + spacing);
+    }
+  } else {
+    int32_t j = 0;
+    int32_t h = 0;
+    int32_t num = 0;
+    int32_t n = children_for_layout->size;
+    uint32_t rows = (n % cols) ? (n / cols) + 1 : (n / cols);
+    for (i = 0; i < rows && num < children_for_layout->size; i++) {
+      for (j = 0, h = 0; j < cols && num < children_for_layout->size; j++, num++) {
+        h = tk_max(h, children[num]->h);
+      }
+      virtual_h += (h + spacing);
+    }
+  }
+  return virtual_h;
+}
+
+static int32_t children_layouter_list_view_for_list_view_get_scroll_view_w(list_view_t* list_view, widget_t* scroll_view, int32_t virtual_h) {
+  int32_t scroll_view_w = 0;
+  widget_t* scroll_bar = NULL;
+  return_value_if_fail(list_view != NULL && scroll_view != NULL, 0);
+  scroll_bar = list_view->scroll_bar;
+  if (list_view->floating_scroll_bar || scroll_bar == NULL || scroll_bar_is_mobile(scroll_bar) ||
+    (!scroll_bar_is_mobile(scroll_bar) && list_view->auto_hide_scroll_bar && scroll_view->h >= virtual_h)) {
+    scroll_view_w = list_view->widget.w;
+  } else {
+    scroll_view_w = list_view->widget.w - scroll_bar->w;
+  }
+
+  return scroll_view_w;
+}
+
+static ret_t children_layouter_list_view_for_list_view_set_scroll_view_info(widget_t* widget, widget_t* scroll_bar, int32_t virtual_h) {
+  scroll_view_t* scroll_view = SCROLL_VIEW(widget);
+  return_value_if_fail(scroll_view != NULL, RET_BAD_PARAMS);
+  
+  if (widget->h >= virtual_h) {
+    scroll_view_set_offset(widget, 0, 0);
+  }
+  if (scroll_bar_is_mobile(scroll_bar)) {
+    scroll_view_set_yslidable(widget, TRUE);
+  }
+
+  scroll_view_set_xslidable(widget, FALSE);
+  scroll_view_set_virtual_h(widget, virtual_h);
+  scroll_view->xoffset = 0;
+  if (scroll_view->yoffset + widget->h > scroll_view->virtual_h) {
+    scroll_view->yoffset = scroll_view->virtual_h - widget->h;
+    scroll_view->yoffset = scroll_view->yoffset > 0 ? scroll_view->yoffset : 0;
+  }
+  if (scroll_view->on_scroll) {
+    scroll_view->on_scroll(widget, scroll_view->xoffset, scroll_view->yoffset);
+  }
+  return RET_OK;
+}
+
+static ret_t children_layouter_list_view_for_list_view_set_scroll_bar_info(widget_t* widget, list_view_t* list_view, widget_t* scroll_view, int32_t virtual_h, int32_t item_height) {
+  return_value_if_fail(widget != NULL && list_view != NULL, RET_BAD_PARAMS);
+  
+  scroll_bar_set_params(widget, virtual_h, item_height);
+  if (scroll_bar_is_mobile(widget)) {
+    widget_set_visible_only(widget, FALSE);
+  } else {
+    if (scroll_view->h >= virtual_h) {
+      scroll_bar_set_value(widget, 0);
+      if (list_view->auto_hide_scroll_bar || list_view->floating_scroll_bar) {
+        widget_set_visible_only(widget, FALSE);
+      } else {
+        widget_set_enable(widget, FALSE);
+        widget_set_visible_only(widget, TRUE);
+      }
+    } else {
+      if (list_view->auto_hide_scroll_bar && list_view->floating_scroll_bar) {
+        widget_set_visible_only(widget, list_view->is_over);
+      } else {
+        widget_set_enable(widget, TRUE);
+        widget_set_visible_only(widget, TRUE);
+      }
+    }
+  }
+  return RET_OK;
+}
+
+static ret_t children_layouter_list_view_for_list_view_layout(children_layouter_t* layouter,
+                                                              widget_t* widget) {
   int32_t virtual_h = 0;
   int32_t item_height = 0;
-  int32_t default_item_height = 0;
   list_view_t* list_view = NULL;
-  widget_t* scroll_bar = NULL;
+  int32_t default_item_height = 0;
   scroll_view_t* scroll_view = SCROLL_VIEW(widget);
   children_layouter_list_view_t* l = (children_layouter_list_view_t*)layouter;
   return_value_if_fail(widget != NULL && scroll_view != NULL && l != NULL, RET_BAD_PARAMS);
 
-  virtual_h = widget->h;
   list_view = LIST_VIEW(widget->parent);
   return_value_if_fail(list_view != NULL, RET_BAD_PARAMS);
 
-  scroll_bar = list_view->scroll_bar;
-  spacing = l->spacing;
-  x_margin = l->x_margin;
-  y_margin = l->y_margin;
   item_height = list_view->item_height ? list_view->item_height : l->item_height;
   default_item_height =
       list_view->default_item_height ? list_view->default_item_height : l->default_item_height;
 
   if (widget->children != NULL) {
-    int32_t i = 0;
-    int32_t n = 0;
-    int32_t x = x_margin;
-    int32_t y = y_margin;
-    int32_t h = item_height;
-    widget_t** children = NULL;
+    int32_t scroll_view_w = 0;
     darray_t children_for_layout;
-    int32_t w = widget->w - x_margin * 2;
     uint32_t cols = l->cols <= 1 ? 1 : l->cols;
 
     widget_layout_floating_children(widget);
@@ -247,132 +392,23 @@ static ret_t children_layouter_for_list_view_layout(children_layouter_t* layoute
                                        l->keep_invisible) == RET_OK,
         RET_BAD_PARAMS);
 
-    n = children_for_layout.size;
-    children = (widget_t**)(children_for_layout.elms);
+    children_layouter_list_view_for_list_view_children_layout_h(&children_for_layout, item_height, default_item_height);
+    virtual_h = children_layouter_list_view_for_list_view_get_virtual_h(&children_for_layout, cols, l->y_margin, l->spacing);
+    scroll_view_w = children_layouter_list_view_for_list_view_get_scroll_view_w(list_view, widget, virtual_h);
 
-    if (cols <= 1) {
-      for (i = 0; i < n; i++) {
-        widget_t* iter = children[i];
-
-        if (item_height <= 0) {
-          h = iter->h;
-        }
-
-        if (h <= 0) {
-          h = default_item_height;
-        }
-
-        y = y + (h > 0 ? h : iter->h) + spacing;
-        if (y > virtual_h) {
-          virtual_h = y;
-        }
-      }
-    } else {
-      uint32_t rows = (n % cols) ? (n / cols) + 1 : (n / cols);
-      if (item_height == 0) item_height = default_item_height;
-      y = (item_height + spacing) * rows;
-      if (y > virtual_h) {
-        virtual_h = y;
-      }
-    }
-
-    scroll_view->widget.w = list_view->widget.w;
-    if (scroll_bar != NULL) {
-      if (!scroll_bar_is_mobile(scroll_bar)) {
-        if (list_view->auto_hide_scroll_bar) {
-          if (virtual_h <= widget->h) {
-            widget_set_visible_only(scroll_bar, FALSE);
-            widget_set_enable(scroll_bar, FALSE);
-          } else {
-            scroll_view->widget.w = list_view->widget.w - scroll_bar->w;
-            widget_set_visible_only(scroll_bar, TRUE);
-            widget_set_enable(scroll_bar, TRUE);
-          }
-        } else {
-          if (!widget_get_prop_bool(scroll_bar, WIDGET_PROP_VISIBLE, FALSE)) {
-            widget_set_visible_only(scroll_bar, TRUE);
-            widget_set_enable(scroll_bar, TRUE);
-          }
-          scroll_view->widget.w = list_view->widget.w - scroll_bar->w;
-        }
-      }
-    }
-
-    y = y_margin;
-    w = scroll_view->widget.w - 2 * x_margin;
-
-    if (cols <= 1) {
-      for (i = 0; i < n; i++) {
-        widget_t* iter = children[i];
-
-        if (item_height <= 0) {
-          h = iter->h;
-        }
-
-        if (h <= 0) {
-          h = default_item_height;
-        }
-
-        widget_move_resize(iter, x, y, w, h);
-        widget_layout(iter);
-
-        y = iter->y + iter->h + spacing;
-      }
-    } else {
-      uint32_t item_w = (w - (cols - 1) * spacing) / cols;
-
-      h = item_height;
-      y = y_margin - item_height - spacing;
-      for (i = 0; i < n; i++) {
-        widget_t* iter = children[i];
-
-        if (i % cols == 0) {
-          x = x_margin;
-          y += item_height + spacing;
-        } else {
-          x += item_w + spacing;
-        }
-
-        widget_move_resize(iter, x, y, item_w, h);
-        widget_layout(iter);
-      }
-    }
-
-    if (scroll_bar != NULL && (SCROLL_BAR(scroll_bar)->value) >= y) {
-      int32_t offset = tk_max(0, (y - widget->h + item_height));
-      scroll_bar_set_value(scroll_bar, offset);
-      scroll_view_set_offset(WIDGET(scroll_view), 0, offset);
-    }
+    widget_resize(widget, scroll_view_w, widget->h);
+    children_layouter_list_view_for_list_view_children_layout_w(&children_for_layout, cols, l->x_margin, l->y_margin, l->spacing, scroll_view_w);
 
     darray_deinit(&(children_for_layout));
-  } else {
-    scroll_bar_set_value(scroll_bar, 0);
-    scroll_view_set_offset(WIDGET(scroll_view), 0, 0);
   }
 
-  scroll_view_set_virtual_h(list_view->scroll_view, virtual_h);
-  item_height = tk_max(item_height, default_item_height);
-  scroll_bar_set_params(list_view->scroll_bar, virtual_h, item_height);
-  scroll_view->xoffset = 0;
-  if (scroll_view->yoffset + widget->h > scroll_view->virtual_h) {
-    scroll_view->yoffset = scroll_view->virtual_h - widget->h;
-    scroll_view->yoffset = scroll_view->yoffset > 0 ? scroll_view->yoffset : 0;
-  }
-  if (scroll_view->on_scroll) {
-    scroll_view->on_scroll(widget, scroll_view->xoffset, scroll_view->yoffset);
-  }
-
-  scroll_view_set_xslidable(list_view->scroll_view, FALSE);
-  if (scroll_bar_is_mobile(list_view->scroll_bar)) {
-    scroll_view_set_yslidable(list_view->scroll_view, TRUE);
-    widget_set_visible_only(list_view->scroll_bar, FALSE);
-  }
-
+  children_layouter_list_view_for_list_view_set_scroll_view_info(widget, list_view->scroll_bar, virtual_h);
+  children_layouter_list_view_for_list_view_set_scroll_bar_info(list_view->scroll_bar, list_view, widget, virtual_h, item_height);
   return RET_OK;
 }
 
-static ret_t children_layouter_for_list_view_h_layout(children_layouter_t* layouter,
-                                                      widget_t* widget) {
+static ret_t children_layouter_list_view_for_list_view_h_layout(children_layouter_t* layouter,
+                                                                widget_t* widget) {
   int32_t virtual_w = 0;
   list_view_h_t* list_view_h = NULL;
   scroll_view_t* scroll_view = SCROLL_VIEW(widget);
@@ -436,9 +472,9 @@ static ret_t children_layouter_list_view_layout(children_layouter_t* layouter, w
   children_layouter_list_view_t* l = (children_layouter_list_view_t*)layouter;
   return_value_if_fail(l != NULL, RET_BAD_PARAMS);
   if (l->hlayouter) {
-    return children_layouter_for_list_view_h_layout(layouter, widget);
+    return children_layouter_list_view_for_list_view_h_layout(layouter, widget);
   } else {
-    return children_layouter_for_list_view_layout(layouter, widget);
+    return children_layouter_list_view_for_list_view_layout(layouter, widget);
   }
 }