+++ src/ext_widgets/scroll_view/scroll_bar.h
@@ -235,7 +235,9 @@ ret_t scroll_bar_set_value_only(widget_t* widget, int32_t value);
  */
 bool_t scroll_bar_is_mobile(widget_t* widget);
 
-ret_t scroll_bar_hide_by_opacity_animation(widget_t* widget, int32_t duration);
+/* private */
+ret_t scroll_bar_hide_by_opacity_animation(widget_t* widget, int32_t duration, int32_t delay);
+ret_t scroll_bar_show_by_opacity_animation(widget_t* widget, int32_t duration, int32_t delay);
 
 #define SCROLL_BAR(widget) ((scroll_bar_t*)(scroll_bar_cast(WIDGET(widget))))